refactor: restructure services and remove unused files; add new DTOs and validation schemas

+++ apps/patient/src/patient.service.ts
@@ -1,14 +1,13 @@
 import { Injectable } from '@nestjs/common';
 import { PrismaService } from '../../../libs/common-utils/src/db/prisma.service';
-import { CreatePatientDto } from './dtos/create-patient.dto';
 import { IUserDecorator } from './types/user';
-import { UpdatePatientDto } from './dtos/update-patient.dto';
+import { ICreatePatientDto, IUpdatePatientDto } from './types/patient';
 
 @Injectable()
 export class PatientService {
     constructor(private readonly prisma: PrismaService) {}
 
-    async create(user: IUserDecorator, createPatientDto: CreatePatientDto) {
+    async create(user: IUserDecorator, createPatientDto: ICreatePatientDto) {
         const { userId } = user;
         return this.prisma.patient.create({
             data: { ...createPatientDto, userId },
@@ -23,7 +22,11 @@ export class PatientService {
         return this.prisma.patient.findUnique({ where: { id } });
     }
 
-    async update(id: number, updatePatientDto: UpdatePatientDto) {
+    async update(id: number, updatePatientDto: IUpdatePatientDto) {
+        const patient = await this.findOne(id);
+        if (!patient) {
+            return null;
+        }
         return this.prisma.patient.update({
             where: { id },
             data: updatePatientDto,
@@ -31,6 +34,10 @@ export class PatientService {
     }
 
     async remove(id: number) {
+        const patient = await this.findOne(id);
+        if (!patient) {
+            return null;
+        }
         return this.prisma.patient.delete({ where: { id } });
     }
 }
